+++ jkube-kit/generator/api/src/main/java/org/eclipse/jkube/generator/api/support/AbstractPortsExtractor.java
@@ -24,12 +24,12 @@ import java.util.Properties;
 import java.util.Set;
 import java.util.regex.Pattern;
 
-import io.fabric8.kubernetes.client.utils.Serialization;
 import org.apache.commons.lang3.StringUtils;
 import org.eclipse.jkube.generator.api.PortsExtractor;
 import org.eclipse.jkube.kit.common.Configs;
 import org.eclipse.jkube.kit.common.JavaProject;
 import org.eclipse.jkube.kit.common.PrefixedLogger;
+import org.eclipse.jkube.kit.common.util.Serialization;
 
 import static org.eclipse.jkube.kit.common.util.PropertiesUtil.toMap;
 
@@ -122,10 +122,8 @@ public abstract class AbstractPortsExtractor implements PortsExtractor {
 	 */
 	private Map<String, String> readConfig(File f) throws IOException {
 		Map<String, String> map;
-		if (f.getName().endsWith(JSON_EXTENSION)) {
-			map = flatten(Serialization.jsonMapper().readValue(f, Map.class));
-		} else if (f.getName().endsWith(YAML_EXTENSION) || f.getName().endsWith(YML_EXTENSION)) {
-			map = flatten(Serialization.yamlMapper().readValue(f, Map.class));
+		if (f.getName().endsWith(JSON_EXTENSION) || f.getName().endsWith(YAML_EXTENSION) || f.getName().endsWith(YML_EXTENSION)) {
+			map = flatten(Serialization.unmarshal(f, Map.class));
 		} else if (f.getName().endsWith(PROPERTIES_EXTENSION)) {
 			Properties properties = new Properties();
 			try (FileInputStream fis = new FileInputStream(f)) {