+++ CHANGELOG.md
@@ -29,6 +29,7 @@ Usage:
 * Fix #2170: `internal-microservice` profile prevents Service exposure
 * Fix #2174: Profile merge constructor accounts for parentProfile field
 * Fix #2187: `serviceAccount` configuration option has stopped working
+* Fix #2192: Bump Kubernetes Client to 6.7.0 (use JKube Serialization util to wrap around the Kubernetes Client KubernetesSerialization)
 
 ### 1.12.0 (2023-04-03)
 * Fix #1179: Move storageClass related functionality out of VolumePermissionEnricher to PersistentVolumeClaimStorageClassEnricher